requirements.txt
ultralytics
streamlit
flask
opencv-python
pyyaml
twilio
python-dotenv
pyobjc